+++ src/HASS.Agent/HASS.Agent.Shared/Managers/Audio/AudioSession.cs
@@ -13,5 +13,5 @@ public class AudioSession
     public bool Muted { get; set; }
     public bool Active { get; set; }
     public int MasterVolume { get; set; }
-    public int PeakVolume { get; set; }
+    public double PeakVolume { get; set; }
 }
